06/16 커밋

+++ domain-module/src/main/java/io/khw/domain/stock/dto/StockIncVolumeDto.java
@@ -15,5 +15,5 @@ public class StockIncVolumeDto {
 
     private String stockCode;
 
-    private double tradeVolume;
+    private int tradeVolume;
 }
